+++ src/application/BlackJack.java
@@ -6,9 +6,33 @@
 
 package application;
 
-public class BlackJack {
+import exceptions.InvalidLogicException;
+
+public class BlackJack extends Table {
 	
-	// Deck
-	// players[]
+	private Deck deck;
+
+	public BlackJack(int gameIndex, int maxPlayers, int ante, String dealer) throws InvalidLogicException 
+	{
+		super(gameIndex, maxPlayers, ante, dealer);
+		// make a deck and shuffle it
+		deck = new Deck();
+		deck.shuffle();
+	}
+	
+	/***
+	 * + mainPlay(): void
+	 * where black jack is played with each hand
+	 * @throws InvalidLogicException 
+	 */
+	public void mainPlay() throws InvalidLogicException
+	{
+		int numOfPlayers = super.players.size();
+		// catch if
+		if(numOfPlayers<1)
+			throw new InvalidLogicException("BlackJack.mainPlay():\tNo players to play"
+					+ "\nplayers:" + players);
+		
+	}
 	
 }